Added dashboard client to receive telemetry messages and display data

+++ inc/devices.h
@@ -78,7 +78,7 @@ public:
     MessageToJsonString(*message, out);
   }
 
-  float timestamp;
+  double timestamp;
 
   int count;
   float helper_variable = 0;
